Make store features additive
