+++ src/stores/sled_store.rs
@@ -1,19 +1,18 @@
-use crate::{Location, StoreImpl};
+use crate::{Location, Store};
 use serde::{de::DeserializeOwned, Serialize};
 
+/// A store implemented using [`sled`]
 #[derive(Debug)]
-pub struct RocksDBStore {
-    db: rocksdb::DB,
+pub struct SledStore {
+    db: sled::Db,
 }
-
-pub use RocksDBStore as InnerStore;
 
 /// Errors that can occur during `PkvStore::get`
 #[derive(thiserror::Error, Debug)]
 pub enum GetError {
-    /// An internal error from the rocksdb crate
-    #[error("Rocksdb error")]
-    Rocksdb(#[from] rocksdb::Error),
+    /// An internal error from the sled crate
+    #[error("Sled error")]
+    Sled(#[from] sled::Error),
     /// Error when deserializing the value
     #[error("MessagePack deserialization error")]
     MessagePack(#[from] rmp_serde::decode::Error),
@@ -25,28 +24,23 @@ pub enum GetError {
 /// Errors that can occur during `PkvStore::set`
 #[derive(thiserror::Error, Debug)]
 pub enum SetError {
-    /// An internal error from the rocksdb crate
-    #[error("Rocksdb error")]
-    Rocksdb(#[from] rocksdb::Error),
+    /// An internal error from the sled crate
+    #[error("Sled error")]
+    Sled(#[from] sled::Error),
     /// Error when serializing the value
     #[error("MessagePack serialization error")]
     MessagePack(#[from] rmp_serde::encode::Error),
 }
 
-impl RocksDBStore {
+impl SledStore {
     pub(crate) fn new(location: Location) -> Self {
-        let mut options = rocksdb::Options::default();
-        options.set_error_if_exists(false);
-        options.create_if_missing(true);
-        options.create_missing_column_families(true);
-
-        let db_path = location.get_path().join("bevy_rocksdb_pkv");
-        let db = rocksdb::DB::open(&options, db_path).expect("Failed to init key value store");
+        let db_path = location.get_path().join("bevy_pkv.sled");
+        let db = sled::open(db_path).expect("Failed to init key value store");
         Self { db }
     }
 }
 
-impl StoreImpl for RocksDBStore {
+impl Store for SledStore {
     type GetError = GetError;
     type SetError = SetError;
 
@@ -54,16 +48,14 @@ impl StoreImpl for RocksDBStore {
     fn set<T: Serialize>(&mut self, key: &str, value: &T) -> Result<(), Self::SetError> {
         let mut serializer = rmp_serde::Serializer::new(Vec::new()).with_struct_map();
         value.serialize(&mut serializer)?;
-        self.db.put(key, serializer.into_inner())?;
-
+        self.db.insert(key, serializer.into_inner())?;
         Ok(())
     }
 
     /// More or less the same as set::<String>, but can take a &str
     fn set_string(&mut self, key: &str, value: &str) -> Result<(), Self::SetError> {
         let bytes = rmp_serde::to_vec(value)?;
-        self.db.put(key, bytes)?;
-
+        self.db.insert(key, bytes)?;
         Ok(())
     }
 
@@ -76,15 +68,9 @@ impl StoreImpl for RocksDBStore {
     }
 
     /// Clear all keys and their values
-    /// The RocksDB adapter uses an iterator to achieve this, unlike sled
+    /// clear is also a kind of store so it will return SetError on failure
     fn clear(&mut self) -> Result<(), Self::SetError> {
-        let kv_iter = self.db.iterator(rocksdb::IteratorMode::Start);
-
-        for kv in kv_iter {
-            let (key, _) = kv?;
-            self.db.delete(key)?;
-        }
-
+        self.db.clear()?;
         Ok(())
     }
 }
